+++ apps/landing/src/app/features/deeplink/constants/index.ts
@@ -1,12 +1,13 @@
 import { storeUrls } from '../../home/constants';
 
 const isLocalhost = window.location.hostname === 'localhost' || window.location.hostname === '127.0.0.1';
-const isDev = import.meta.env.DEV || window.location.hostname.includes('-dev');
+const isDev =
+    import.meta.env.DEV || window.location.hostname.includes('-dev') || window.location.hostname.includes('app-dev');
 
 export const APP_CONFIG = {
     name: 'DoU',
-    scheme: 'chatic',
-    packageId: 'io.chatic.dou',
+    scheme: isDev ? 'chatic-dev' : 'chatic',
+    packageId: isDev ? 'io.chatic.dou.dev' : 'io.chatic.dou',
     appStoreId: '6758658673',
     storeUrls,
 } as const;